Remove circular import common <-> test module
To remove circular import, this moves modules/test/helpers into
common/test/helpers.
Also integrates the modules/test-operators into modules/test and
removes the need for the modules/test-operators module.
