--- modules/test-operators/helpers/common.go
@@ -1,44 +1,0 @@
-/*
-Copyright 2023 Red Hat
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package helpers
-
-import (
-	"context"
-	"time"
-
-	"github.com/go-logr/logr"
-
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	base "github.com/openstack-k8s-operators/lib-common/modules/test/helpers"
-)
-
-// TestHelper is a collection of helpers for testing operators. It extends the
-// generic TestHelper from modules/test.
-type TestHelper struct {
-	*base.TestHelper
-}
-
-// NewTestHelper returns a TestHelper
-func NewTestHelper(
-	ctx context.Context,
-	k8sClient client.Client,
-	timeout time.Duration,
-	interval time.Duration,
-	logger logr.Logger,
-) *TestHelper {
-	helper := &TestHelper{}
-	helper.TestHelper = base.NewTestHelper(ctx, k8sClient, timeout, interval, logger)
-	return helper
-}
